Bring Nelson's build to the modern age
* Bump to SBT 1.0.4
* Update plugins
* Remove WartRemover - too many triggered, don't think it gives much
  benefit for the pain of continuing to use it
* Delete ALLLL the unused imports omg

+++ core/src/main/scala/Cache.scala
@@ -95,7 +95,7 @@ class GuavaCacheBuilder[K, V] extends CacheBuilder[K, V] {
   def build(): Cache[K, V] = new GuavaCache[K, V](builder.build[K, V]())
 
   def maximumSize(n: Int) = new GuavaCacheBuilder[K, V] {
-    override val builder = GuavaCacheBuilder.this.builder.maximumSize(n)
+    override val builder = GuavaCacheBuilder.this.builder.maximumSize(n.toLong)
   }
 
   def concurrencyLevel(n: Int) = new GuavaCacheBuilder[K, V] {
